+++ exec/plugin-commands-script-runners/test/index.ts
@@ -1,5 +1,5 @@
 /// <reference path="../../../__typings__/index.d.ts" />
-import { promises as fs } from 'fs'
+import fsSync, { promises as fs } from 'fs'
 import path from 'path'
 import { type PnpmError } from '@pnpm/error'
 import { readProjects } from '@pnpm/filter-workspace-packages'
@@ -9,6 +9,7 @@ import {
   test as testCommand,
 } from '@pnpm/plugin-commands-script-runners'
 import { prepare, preparePackages } from '@pnpm/prepare'
+import { createTestIpcServer } from '@pnpm/test-ipc-server'
 import execa from 'execa'
 import isWindows from 'is-windows'
 import writeYamlFile from 'write-yaml-file'
@@ -179,23 +180,24 @@ test('run stop: pass the args to the command that is specified in the build scri
 })
 
 test('restart: run stop, restart and start', async () => {
+  await using server = await createTestIpcServer()
+
   prepare({
     scripts: {
-      poststop: 'node -e "process.stdout.write(\'poststop\')" | json-append ./output.json',
-      prestop: 'node -e "process.stdout.write(\'prestop\')" | json-append ./output.json',
-      stop: 'node -e "process.stdout.write(\'stop\')" | json-append ./output.json',
+      poststop: server.sendLineScript('poststop'),
+      prestop: server.sendLineScript('prestop'),
+      stop: server.sendLineScript('stop'),
 
-      postrestart: 'node -e "process.stdout.write(\'postrestart\')" | json-append ./output.json',
-      prerestart: 'node -e "process.stdout.write(\'prerestart\')" | json-append ./output.json',
-      restart: 'node -e "process.stdout.write(\'restart\')" | json-append ./output.json',
+      postrestart: server.sendLineScript('postrestart'),
+      prerestart: server.sendLineScript('prerestart'),
+      restart: server.sendLineScript('restart'),
 
-      poststart: 'node -e "process.stdout.write(\'poststart\')" | json-append ./output.json',
-      prestart: 'node -e "process.stdout.write(\'prestart\')" | json-append ./output.json',
-      start: 'node -e "process.stdout.write(\'start\')" | json-append ./output.json',
+      poststart: server.sendLineScript('poststart'),
+      prestart: server.sendLineScript('prestart'),
+      start: server.sendLineScript('start'),
     },
   })
 
-  await execa('pnpm', ['add', 'json-append@1'])
   await restart.handler({
     dir: process.cwd(),
     extraBinPaths: [],
@@ -203,8 +205,7 @@ test('restart: run stop, restart and start', async () => {
     rawConfig: {},
   }, [])
 
-  const { default: scriptsRan } = await import(path.resolve('output.json'))
-  expect(scriptsRan).toStrictEqual([
+  expect(server.getLines()).toStrictEqual([
     'stop',
     'restart',
     'start',
@@ -212,23 +213,24 @@ test('restart: run stop, restart and start', async () => {
 })
 
 test('restart: run stop, restart and start and all the pre/post scripts', async () => {
+  await using server = await createTestIpcServer()
+
   prepare({
     scripts: {
-      poststop: 'node -e "process.stdout.write(\'poststop\')" | json-append ./output.json',
-      prestop: 'node -e "process.stdout.write(\'prestop\')" | json-append ./output.json',
-      stop: 'pnpm prestop && node -e "process.stdout.write(\'stop\')" | json-append ./output.json && pnpm poststop',
+      poststop: server.sendLineScript('poststop'),
+      prestop: server.sendLineScript('prestop'),
+      stop: `${server.sendLineScript('stop')} && pnpm poststop`,
 
-      postrestart: 'node -e "process.stdout.write(\'postrestart\')" | json-append ./output.json',
-      prerestart: 'node -e "process.stdout.write(\'prerestart\')" | json-append ./output.json',
-      restart: 'node -e "process.stdout.write(\'restart\')" | json-append ./output.json',
+      postrestart: server.sendLineScript('postrestart'),
+      prerestart: server.sendLineScript('prerestart'),
+      restart: server.sendLineScript('restart'),
 
-      poststart: 'node -e "process.stdout.write(\'poststart\')" | json-append ./output.json',
-      prestart: 'node -e "process.stdout.write(\'prestart\')" | json-append ./output.json',
-      start: 'node -e "process.stdout.write(\'start\')" | json-append ./output.json',
+      poststart: server.sendLineScript('poststart'),
+      prestart: server.sendLineScript('prestart'),
+      start: server.sendLineScript('start'),
     },
   })
 
-  await execa('pnpm', ['add', 'json-append@1'])
   await restart.handler({
     dir: process.cwd(),
     enablePrePostScripts: true,
@@ -237,8 +239,7 @@ test('restart: run stop, restart and start and all the pre/post scripts', async 
     rawConfig: {},
   }, [])
 
-  const { default: scriptsRan } = await import(path.resolve('output.json'))
-  expect(scriptsRan).toStrictEqual([
+  expect(server.getLines()).toStrictEqual([
     'prestop',
     'stop',
     'poststop',
@@ -281,9 +282,6 @@ test('"pnpm run" prints the list of available commands, including commands of th
     {
       location: '.',
       package: {
-        dependencies: {
-          'json-append': '1',
-        },
         scripts: {
           build: 'echo root',
           test: 'test-all',
@@ -372,11 +370,8 @@ test('if a script is not found but is present in the root, print an info message
     {
       location: '.',
       package: {
-        dependencies: {
-          'json-append': '1',
-        },
         scripts: {
-          build: 'node -e "process.stdout.write(\'root\')" | json-append ./output.json',
+          build: 'node -e "process.stdout.write(\'root\')"',
         },
       },
     },
@@ -417,12 +412,13 @@ test('if a script is not found but is present in the root, print an info message
 test('scripts work with PnP', async () => {
   prepare({
     scripts: {
-      foo: 'node -e "process.stdout.write(\'foo\')" | json-append ./output.json',
+      foo: 'hello-world-js-bin > ./output.txt',
     },
   })
 
-  await execa(pnpmBin, ['add', 'json-append@1'], {
+  await execa(pnpmBin, ['add', '@pnpm.e2e/hello-world-js-bin@1.0.0'], {
     env: {
+      NPM_CONFIG_REGISTRY: REGISTRY_URL,
       NPM_CONFIG_NODE_LINKER: 'pnp',
       NPM_CONFIG_SYMLINK: 'false',
     },
@@ -434,8 +430,11 @@ test('scripts work with PnP', async () => {
     rawConfig: {},
   }, ['foo'])
 
-  const { default: scriptsRan } = await import(path.resolve('output.json'))
-  expect(scriptsRan).toStrictEqual(['foo'])
+  // https://github.com/pnpm/registry-mock/blob/ac2e129eb262009d2e7cd43ed869c31097793073/packages/hello-world-js-bin%401.0.0/index.js#L2
+  const helloWorldJsBinOutput = 'Hello world!\n'
+
+  const fooOutput = fsSync.readFileSync(path.resolve('output.txt')).toString()
+  expect(fooOutput).toStrictEqual(helloWorldJsBinOutput)
 })
 
 test('pnpm run with custom shell', async () => {
@@ -517,14 +516,15 @@ test('pnpm run with RegExp script selector should work also for pre/post script'
 })
 
 test('pnpm run with RegExp script selector should work parallel as a default behavior (parallel execution limits number is four)', async () => {
+  await using serverA = await createTestIpcServer()
+  await using serverB = await createTestIpcServer()
+
   prepare({
     scripts: {
-      'build:a': 'node -e "let i = 20;setInterval(() => {if (!--i) process.exit(0); require(\'json-append\').append(Date.now(),\'./output-a.json\');},50)"',
-      'build:b': 'node -e "let i = 40;setInterval(() => {if (!--i) process.exit(0); require(\'json-append\').append(Date.now(),\'./output-b.json\');},25)"',
+      'build:a': `node -e "let i = 20;setInterval(() => {if (!--i) process.exit(0); console.log(Date.now());},50)" | ${serverA.generateSendStdinScript()}`,
+      'build:b': `node -e "let i = 40;setInterval(() => {if (!--i) process.exit(0); console.log(Date.now());},25)" | ${serverB.generateSendStdinScript()}`,
     },
   })
-
-  await execa('pnpm', ['add', 'json-append@1'])
 
   await run.handler({
     dir: process.cwd(),
@@ -533,21 +533,22 @@ test('pnpm run with RegExp script selector should work parallel as a default beh
     rawConfig: {},
   }, ['/build:.*/'])
 
-  const { default: outputsA } = await import(path.resolve('output-a.json'))
-  const { default: outputsB } = await import(path.resolve('output-b.json'))
+  const outputsA = serverA.getLines().map(x => Number.parseInt(x))
+  const outputsB = serverB.getLines().map(x => Number.parseInt(x))
 
   expect(Math.max(outputsA[0], outputsB[0]) < Math.min(outputsA[outputsA.length - 1], outputsB[outputsB.length - 1])).toBeTruthy()
 })
 
 test('pnpm run with RegExp script selector should work sequentially with --workspace-concurrency=1', async () => {
+  await using serverA = await createTestIpcServer()
+  await using serverB = await createTestIpcServer()
+
   prepare({
     scripts: {
-      'build:a': 'node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); require(\'json-append\').append(Date.now(),\'./output-a.json\');},16)"',
-      'build:b': 'node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); require(\'json-append\').append(Date.now(),\'./output-b.json\');},16)"',
+      'build:a': `node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); console.log(Date.now()); },16)" | ${serverA.generateSendStdinScript()}`,
+      'build:b': `node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); console.log(Date.now()); },16)" | ${serverB.generateSendStdinScript()}`,
     },
   })
-
-  await execa('pnpm', ['add', 'json-append@1'])
 
   await run.handler({
     dir: process.cwd(),
@@ -557,17 +558,20 @@ test('pnpm run with RegExp script selector should work sequentially with --works
     workspaceConcurrency: 1,
   }, ['/build:.*/'])
 
-  const { default: outputsA } = await import(path.resolve('output-a.json'))
-  const { default: outputsB } = await import(path.resolve('output-b.json'))
+  const outputsA = serverA.getLines().map(x => Number.parseInt(x))
+  const outputsB = serverB.getLines().map(x => Number.parseInt(x))
 
   expect(outputsA[0] < outputsB[0] && outputsA[1] < outputsB[1]).toBeTruthy()
 })
 
 test('pnpm run with RegExp script selector with flag should throw error', async () => {
+  await using serverA = await createTestIpcServer()
+  await using serverB = await createTestIpcServer()
+
   prepare({
     scripts: {
-      'build:a': 'node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); require(\'json-append\').append(Date.now(),\'./output-a.json\');},16)"',
-      'build:b': 'node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); require(\'json-append\').append(Date.now(),\'./output-b.json\');},16)"',
+      'build:a': `node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); console.log(Date.now()); },16)" | ${serverA.generateSendStdinScript()}`,
+      'build:b': `node -e "let i = 2;setInterval(() => {if (!i--) process.exit(0); console.log(Date.now()); },16)" | ${serverB.generateSendStdinScript()}`,
     },
   })