+++ exec/prepare-package/test/index.ts
@@ -1,25 +1,27 @@
 import path from 'path'
 import { preparePackage } from '@pnpm/prepare-package'
 import { tempDir } from '@pnpm/prepare'
+import { createTestIpcServer } from '@pnpm/test-ipc-server'
 import { fixtures } from '@pnpm/test-fixtures'
-import { sync as loadJsonFile } from 'load-json-file'
 
 const f = fixtures(__dirname)
 
 test('prepare package runs the prepublish script', async () => {
   const tmp = tempDir()
+  await using server = await createTestIpcServer(path.join(tmp, 'test.sock'))
   f.copy('has-prepublish-script', tmp)
   await preparePackage({ rawConfig: {} }, tmp)
-  expect(loadJsonFile(path.join(tmp, 'output.json'))).toStrictEqual([
+  expect(server.getLines()).toStrictEqual([
     'prepublish',
   ])
 })
 
 test('prepare package does not run the prepublish script if the main file is present', async () => {
   const tmp = tempDir()
+  await using server = await createTestIpcServer(path.join(tmp, 'test.sock'))
   f.copy('has-prepublish-script-and-main-file', tmp)
   await preparePackage({ rawConfig: {} }, tmp)
-  expect(loadJsonFile(path.join(tmp, 'output.json'))).toStrictEqual([
+  expect(server.getLines()).toStrictEqual([
     'prepublish',
   ])
 })