+++ env/node.resolver/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/node.resolver
 
+## 2.0.38
+
+### Patch Changes
+
+- @pnpm/node.fetcher@3.0.37
+
 ## 2.0.37
 
 ### Patch Changes