+++ exec/plugin-commands-rebuild/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @pnpm/plugin-commands-rebuild
 
+## 10.0.14
+
+### Patch Changes
+
+- @pnpm/worker@0.3.13
+- @pnpm/store-connection-manager@7.0.23
+
+## 10.0.13
+
+### Patch Changes
+
+- Updated dependencies [c597f72ec]
+  - @pnpm/config@20.4.0
+  - @pnpm/cli-utils@2.1.7
+  - @pnpm/store-connection-manager@7.0.22
+  - @pnpm/workspace.find-packages@1.1.8
+
 ## 10.0.12
 
 ### Patch Changes