+++ store/plugin-commands-server/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pnpm/plugin-commands-server
 
+## 6.0.54
+
+### Patch Changes
+
+- @pnpm/server@17.0.6
+- @pnpm/store-connection-manager@7.0.23
+
+## 6.0.53
+
+### Patch Changes
+
+- Updated dependencies [c597f72ec]
+  - @pnpm/config@20.4.0
+  - @pnpm/cli-utils@2.1.7
+  - @pnpm/store-connection-manager@7.0.22
+
 ## 6.0.52
 
 ### Patch Changes